+++ terraform/modules/github-actions/versions.tf
@@ -1,0 +1,5 @@
+terraform {
+  required_providers {
+    github = { source = "integrations/github", version = "~> 6.0" }
+  }
+}