+++ terraform/providers.tf
@@ -1,0 +1,6 @@
+provider "github" {
+  owner = coalesce(var.github_owner, "disabled")
+  token = var.github_token
+}
+
+provider "sops" {}