+++ terraform/versions.tf
@@ -1,0 +1,7 @@
+terraform {
+  required_version = ">= 1.5.0"
+  required_providers {
+    github = { source = "integrations/github", version = "~> 6.0" }
+    sops   = { source = "carlpett/sops", version = "~> 1.0" }
+  }
+}